+++ Z3Wrap/Expressions/Functions/LambdaExprExtensions.cs
@@ -1,5 +1,3 @@
-using Spaceorc.Z3Wrap.Expressions.Common;
-
 namespace Spaceorc.Z3Wrap.Expressions.Functions;
 
 /// <summary>